feat: dev서버용 master 로그인 api 생성
